Extract SuppressionHandler
