+++ analyzers/tests/SonarAnalyzer.Test/AnalysisContext/SonarAnalysisContextTest.cs
@@ -232,7 +232,7 @@ public partial class SonarAnalysisContextTest
                 {
                     // Verifier expects all diagnostics to increase the counter in order to check that all rules call the
                     // extension method and not the direct `ReportDiagnostic`.
-                    DiagnosticVerifier.SuppressionHandler.IncrementReportCount(context.Diagnostic.Id);
+                    SuppressionHandler.IncrementReportCount(context.Diagnostic.Id);
                     context.ReportDiagnostic(context.Diagnostic);
                 }
             };
